fix:fix git problems

+++ code/backend/weibo-service/src/main/java/com/ilife/weiboservice/dao/UserDao.java
@@ -3,12 +3,12 @@ package com.ilife.weiboservice.dao;
 import com.ilife.weiboservice.entity.User;
 
 
-public interface UserDao{
-    User findAllById(Integer id);
+public interface UserDao {
+    User findAllById(Long id);
 
     User findByNickname(String nickname);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 
     User save(User user);
